% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Algorithm to find the shortest path from an initial node to all other nodes in a graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{data.frame with three variables (v1, v2 and w) that contains the 
edges of the graph (from v1 to v2) with the weight of the edge (w).}

\item{init_node}{A number corresponding to the start node}
}
\value{
The shortest path to every other node from the starting node as a vector.
}
\description{
The algorithm takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.
}
\examples{
wiki_graph <- Lab3dabk::wiki_graph
dijkstra(wiki_graph, 1)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Dijkstra’s_algorithm}
}
